Moved FoFGroups into a new group_identification subpackage
